endpoint filesystem standardisation, unnecessary public funcs privatised commit 1
