+++ internal/endpoints/accountgroups/accountgroups_crud.go
@@ -1,0 +1,140 @@
+package accountgroups
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// ResourceJamfProAccountGroupCreate is responsible for creating a new Jamf Pro Script in the remote system.
+func ResourceJamfProAccountGroupCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+
+	resource, err := constructJamfProAccountGroup(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Account Group: %v", err))
+	}
+
+	var creationResponse *jamfpro.ResponseAccountGroupCreated
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutCreate), func() *retry.RetryError {
+		var apiErr error
+		creationResponse, apiErr = client.CreateAccountGroup(resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return append(diags, diag.FromErr(fmt.Errorf("failed to create Jamf Pro Account Group '%s' after retries: %v", resource.Name, err))...)
+	}
+
+	d.SetId(strconv.Itoa(creationResponse.ID))
+
+	// region concurrency
+	// checkResourceExists := func(id interface{}) (interface{}, error) {
+	// 	intID, err := strconv.Atoi(id.(string))
+	// 	if err != nil {
+	// 		return nil, fmt.Errorf("error converting ID '%v' to integer: %v", id, err)
+	// 	}
+	// 	return client.GetAccountGroupByID(intID)
+	// }
+	// _, waitDiags := waitfor.ResourceIsAvailable(ctx, d, "Jamf Pro Account Group", strconv.Itoa(creationResponse.ID), checkResourceExists, time.Duration(common.DefaultPropagationTime)*time.Second)
+	// if waitDiags.HasError() {
+	// 	return waitDiags
+	// }
+	// endregion
+
+	return append(diags, ResourceJamfProAccountGroupRead(ctx, d, meta)...)
+}
+
+// ResourceJamfProAccountGroupRead is responsible for reading the current state of a Jamf Pro Account Group Resource from the remote system.
+func ResourceJamfProAccountGroupRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := client.GetAccountGroupByID(resourceIDInt)
+
+	if err != nil {
+		return state.HandleResourceNotFoundError(err, d)
+	}
+
+	return updateTerraformState(d, resource)
+}
+
+// ResourceJamfProAccountGroupUpdate is responsible for updating an existing Jamf Pro Account Group on the remote system.
+func ResourceJamfProAccountGroupUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := constructJamfProAccountGroup(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Account Group for update: %v", err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutUpdate), func() *retry.RetryError {
+		_, apiErr := client.UpdateAccountGroupByID(resourceIDInt, resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+
+		return nil
+	})
+
+	// TODO should this be a program breakpoint or just a warn?
+	if err != nil {
+		return append(diags, diag.FromErr(fmt.Errorf("failed to update Jamf Pro Account Group '%s' (ID: %s) after retries: %v", resource.Name, resourceID, err))...)
+	}
+
+	return append(diags, ResourceJamfProAccountGroupRead(ctx, d, meta)...)
+}
+
+// ResourceJamfProAccountGroupDelete is responsible for deleting a Jamf Pro account group.
+func ResourceJamfProAccountGroupDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutDelete), func() *retry.RetryError {
+		apiErr := client.DeleteAccountGroupByID(resourceIDInt)
+		if apiErr != nil {
+			resourceName := d.Get("name").(string)
+			apiErrByName := client.DeleteAccountGroupByName(resourceName)
+			if apiErrByName != nil {
+				return retry.RetryableError(apiErrByName)
+			}
+		}
+		return nil
+	})
+
+	if err != nil {
+		diags = append(diags, diag.FromErr(fmt.Errorf("failed to delete Jamf Pro Account Group '%s' (ID: %s) after retries: %v", d.Get("name").(string), resourceID, err))...)
+	}
+
+	d.SetId("")
+
+	return diags
+}
