+++ internal/endpoints/accounts/accounts_crud.go
@@ -1,0 +1,143 @@
+package accounts
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// ResourceJamfProAccountCreate is responsible for creating a new Jamf Pro Script in the remote system.
+func resourceJamfProAccountCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+
+	var diags diag.Diagnostics
+
+	resource, err := constructJamfProAccount(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Account: %v", err))
+	}
+
+	var creationResponse *jamfpro.ResponseAccountCreatedAndUpdated
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutCreate), func() *retry.RetryError {
+		var apiErr error
+		creationResponse, apiErr = client.CreateAccount(resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to create Jamf Pro Account '%s' after retries: %v", resource.Name, err))
+	}
+
+	d.SetId(strconv.Itoa(creationResponse.ID))
+
+	// checkResourceExists := func(id interface{}) (interface{}, error) {
+	// 	intID, err := strconv.Atoi(id.(string))
+	// 	if err != nil {
+	// 		return nil, fmt.Errorf("error converting ID '%v' to integer: %v", id, err)
+	// 	}
+	// 	return client.GetAccountByID(intID)
+	// }
+
+	// _, waitDiags := waitfor.ResourceIsAvailable(ctx, d, "Jamf Pro Account", strconv.Itoa(creationResponse.ID), checkResourceExists, time.Duration(common.DefaultPropagationTime)*time.Second)
+	// if waitDiags.HasError() {
+	// 	return waitDiags
+	// }
+
+	return append(diags, resourceJamfProAccountRead(ctx, d, meta)...)
+}
+
+// ResourceJamfProAccountRead is responsible for reading the current state of a Jamf Pro Account Group Resource from the remote system.
+func resourceJamfProAccountRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := client.GetAccountByID(resourceIDInt)
+
+	if err != nil {
+		return state.HandleResourceNotFoundError(err, d)
+	}
+
+	return append(diags, updateTerraformState(d, resource)...)
+}
+
+// ResourceJamfProAccountUpdate is responsible for updating an existing Jamf Pro Account Group on the remote system.
+func resourceJamfProAccountUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := constructJamfProAccount(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Account for update: %v", err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutUpdate), func() *retry.RetryError {
+		_, apiErr := client.UpdateAccountByID(resourceIDInt, resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to update Jamf Pro Account '%s' (ID: %s) after retries: %v", resource.Name, resourceID, err))
+	}
+
+	return append(diags, resourceJamfProAccountRead(ctx, d, meta)...)
+}
+
+// ResourceJamfProAccountDelete is responsible for deleting a Jamf Pro account .
+func resourceJamfProAccountDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutDelete), func() *retry.RetryError {
+		apiErr := client.DeleteAccountByID(resourceIDInt)
+		if apiErr != nil {
+			resourceName := d.Get("name").(string)
+			apiErrByName := client.DeleteAccountByName(resourceName)
+			if apiErrByName != nil {
+
+				return retry.RetryableError(apiErrByName)
+			}
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to delete Jamf Pro Account '%s' (ID: %s) after retries: %v", d.Get("name").(string), resourceID, err))
+	}
+
+	d.SetId("")
+
+	return diags
+}
