+++ internal/endpoints/advancedcomputersearches/advancedcomputersearches_constructor.go
