+++ internal/endpoints/advancedmobiledevicesearches/advancedmobiledevicesearches_crud.go
@@ -1,0 +1,140 @@
+package advancedmobiledevicesearches
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// resourceJamfProAdvancedMobileDeviceSearchCreate is responsible for creating a new Jamf Pro mobile device Search in the remote system.
+func resourceJamfProAdvancedMobileDeviceSearchCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+
+	resource, err := constructJamfProAdvancedMobileDeviceSearch(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Advanced Mobile Device Search: %v", err))
+	}
+
+	var creationResponse *jamfpro.ResponseAdvancedMobileDeviceSearchCreatedAndUpdated
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutCreate), func() *retry.RetryError {
+		var apiErr error
+		creationResponse, apiErr = client.CreateAdvancedMobileDeviceSearch(resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to create Jamf Pro Advanced Mobile Device Search '%s' after retries: %v", resource.Name, err))
+	}
+
+	d.SetId(strconv.Itoa(creationResponse.ID))
+
+	// checkResourceExists := func(id interface{}) (interface{}, error) {
+	// 	intID, err := strconv.Atoi(id.(string))
+	// 	if err != nil {
+	// 		return nil, fmt.Errorf("error converting ID '%v' to integer: %v", id, err)
+	// 	}
+	// 	return client.GetAdvancedComputerSearchByID(intID)
+	// }
+
+	// _, waitDiags := waitfor.ResourceIsAvailable(ctx, d, "Jamf Pro Advanced Mobile Device Search", strconv.Itoa(creationResponse.ID), checkResourceExists, time.Duration(common.DefaultPropagationTime)*time.Second)
+	// if waitDiags.HasError() {
+	// 	return waitDiags
+	// }
+
+	return append(diags, resourceJamfProAdvancedMobileDeviceSearchRead(ctx, d, meta)...)
+}
+
+// resourceJamfProAdvancedMobileDeviceSearchRead is responsible for reading the current state of a Jamf Pro mobile device Search from the remote system.
+func resourceJamfProAdvancedMobileDeviceSearchRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := client.GetAdvancedMobileDeviceSearchByID(resourceIDInt)
+	if err != nil {
+		return state.HandleResourceNotFoundError(err, d)
+	}
+
+	return append(diags, updateTerraformState(d, resource)...)
+}
+
+// resourceJamfProAdvancedMobileDeviceSearchUpdate is responsible for updating an existing Jamf Pro mobile device Search on the remote system.
+func resourceJamfProAdvancedMobileDeviceSearchUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := constructJamfProAdvancedMobileDeviceSearch(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Advanced Mobile Device Search for update: %v", err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutUpdate), func() *retry.RetryError {
+		_, apiErr := client.UpdateAdvancedMobileDeviceSearchByID(resourceIDInt, resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to update Jamf Pro Advanced Mobile Device Search '%s' (ID: %s) after retries: %v", resource.Name, resourceID, err))
+	}
+
+	return append(diags, resourceJamfProAdvancedMobileDeviceSearchRead(ctx, d, meta)...)
+}
+
+// resourceJamfProAdvancedMobileDeviceSearchDelete is responsible for deleting a Jamf Pro AdvancedMobileDeviceSearch.
+func resourceJamfProAdvancedMobileDeviceSearchDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutDelete), func() *retry.RetryError {
+		apiErr := client.DeleteAdvancedMobileDeviceSearchByID(resourceIDInt)
+		if apiErr != nil {
+			resourceName := d.Get("name").(string)
+			apiErrByName := client.DeleteAdvancedMobileDeviceSearchByName(resourceName)
+			if apiErrByName != nil {
+				return retry.RetryableError(apiErrByName)
+			}
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to delete Jamf Pro Advanced Mobile Device Search '%s' (ID: %s) after retries: %v", d.Get("name").(string), resourceID, err))
+	}
+
+	d.SetId("")
+
+	return diags
+}
