+++ internal/endpoints/advancedmobiledevicesearches/advancedmobiledevicesearches_stater.go
