+++ internal/endpoints/allowedfileextensions/allowedfileextensions_stater.go
