+++ internal/endpoints/buildings/buildings_crud.go
@@ -1,0 +1,127 @@
+package buildings
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// resourceJamfProBuildingCreate is responsible for creating a new Building in the remote system.
+// The function:
+// 1. Constructs the building data using the provided Terraform configuration.
+// 2. Calls the API to create the building in Jamf Pro.
+// 3. Updates the Terraform state with the ID of the newly created building.
+// 4. Initiates a read operation to synchronize the Terraform state with the actual state in Jamf Pro.
+func resourceJamfProBuildingCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+
+	resource, err := constructJamfProBuilding(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Building: %v", err))
+	}
+
+	var creationResponse *jamfpro.ResponseBuildingCreate
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutCreate), func() *retry.RetryError {
+		var apiErr error
+		creationResponse, apiErr = client.CreateBuilding(resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to create Jamf Pro Building '%s' after retries: %v", resource.Name, err))
+	}
+
+	d.SetId(creationResponse.ID)
+
+	// checkResourceExists := func(id interface{}) (interface{}, error) {
+	// 	return client.GetBuildingByID(id.(string))
+	// }
+
+	// _, waitDiags := waitfor.ResourceIsAvailable(ctx, d, "Jamf Pro Building", creationResponse.ID, checkResourceExists, time.Duration(common.DefaultPropagationTime)*time.Second)
+
+	// if waitDiags.HasError() {
+	// 	return waitDiags
+	// }
+
+	return append(diags, resourceJamfProBuildingRead(ctx, d, meta)...)
+}
+
+// resourceJamfProBuildingRead is responsible for reading the current state of a Building Resource from the remote system.
+// The function:
+// 1. Fetches the building's current state using its ID. If it fails, then obtain the building's current state using its Name.
+// 2. Updates the Terraform state with the fetched data to ensure it accurately reflects the current state in Jamf Pro.
+// 3. Handles any discrepancies, such as the building being deleted outside of Terraform, to keep the Terraform state synchronized.
+func resourceJamfProBuildingRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resource, err := client.GetBuildingByID(resourceID)
+	if err != nil {
+		return state.HandleResourceNotFoundError(err, d)
+	}
+
+	return append(diags, updateTerraformState(d, resource)...)
+}
+
+// resourceJamfProBuildingUpdate is responsible for updating an existing Building on the remote system.
+func resourceJamfProBuildingUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resource, err := constructJamfProBuilding(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Building for update: %v", err))
+	}
+
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutUpdate), func() *retry.RetryError {
+		_, apiErr := client.UpdateBuildingByID(resourceID, resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to update Jamf Pro Building '%s' (ID: %s) after retries: %v", resource.Name, resourceID, err))
+	}
+
+	return append(diags, resourceJamfProBuildingRead(ctx, d, meta)...)
+}
+
+// resourceJamfProBuildingDelete is responsible for deleting a Building.
+func resourceJamfProBuildingDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	err := retry.RetryContext(ctx, d.Timeout(schema.TimeoutDelete), func() *retry.RetryError {
+		apiErr := client.DeleteBuildingByID(resourceID)
+		if apiErr != nil {
+			resourceName := d.Get("name").(string)
+			apiErrByName := client.DeleteBuildingByName(resourceName)
+			if apiErrByName != nil {
+				return retry.RetryableError(apiErrByName)
+			}
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to delete Jamf Pro Building '%s' (ID: %s) after retries: %v", d.Get("name").(string), resourceID, err))
+	}
+
+	d.SetId("")
+
+	return diags
+}
