+++ internal/endpoints/buildings/buildings_stater.go
