+++ internal/endpoints/categories/categories_constructor.go
