+++ internal/endpoints/categories/categories_stater.go
