+++ internal/endpoints/computercheckin/computercheckin_constructor.go
